Update HotbarPets.java
I have implemented two new pets to this file: Ghast pet and Patriot pet. The Patriot Pet is dedicated to myself for continuing this project, and the Ghast pet is brand new, which allows you to shoot fireballs while gaining Fire Resistance. I have also tweaked the imports and most of the formatting. Thank you for allowing me to continue this project Busy!